refactor: 调整当前用户处理
